Split tracking fix (near end issues)
Addresses issues in: Split_tracking.py errors if S + E are too close to end of video #34

Now added in so video doesn't crash when you scroll to end. You need to have at least 5 frames to make a background, otherwise it will NaN the rest of the video. Also, in this case,if you have asked to retrack the next video it will use divide_tracking with 60min.